update Err card user

+++ controllers/cards.js
@@ -1,55 +1,71 @@
 const ObjectId = require('mongoose').Types.ObjectId;
 
 const Card = require('../models/card.js');
+const { CodeSuccess, CodeError } = require('../constants');
 
 module.exports.getAllCards = (req, res) => {
   Card.find({})
-    .then((cards) => res.status(200).send({ data: cards }))
-    .catch((err) => ((err.name === 'ValidationError') ? res.status(400).send({ message: 'Переданы некорректные данные при создании карточки' }) : res.status(500).send({ message: 'Произошла ошибка' })));
+    .then((cards) => res.status(CodeSuccess.OK).send({ data: cards }))
+    .catch((err) => ((err.name === 'ValidationError')
+      ? res.status(CodeError.BAD_REQUEST).send({ message: 'Переданы некорректные данные при создании карточки' })
+      : res.status(CodeError.SERVER_ERROR).send({ message: 'Произошла ошибка' })));
 };
 
 module.exports.createCard = (req, res) => {
   const { name, link } = req.body;
   Card.create({ name, link, owner: req.user._id })
-    .then((card) => res.status(200).send({ data: card }))
-    .catch(() => res.status(500).send({ message: 'Произошла ошибка' }));
+    .then((card) => res.status(CodeSuccess.CREATED).send({ data: card }))
+    .catch((err) => ((err.name === 'CastError' || err.name === 'ValidationError')
+      ? res.status(CodeError.BAD_REQUEST).send({ message: 'Переданы некорректные данные при создании карточки' })
+      : res.status(CodeError.SERVER_ERROR).send({ message: 'Произошла ошибка' })));
 };
 
 module.exports.deleteCard = (req, res) => {
-
+  if (!ObjectId.isValid(req.params)) {
+    res.status(CodeError.BAD_REQUEST).send({ message: 'Передан невалидный id' });
+    return;
+  }
   Card.findById(req.params.id)
     .then((card) => {
       if (!card) {
-        res.status(404).send({ message: 'Карточка с указанным _id не найдена' });
+        res.status(CodeError.NOT_FOUND).send({ message: 'Карточка с указанным _id не найдена' });
       } else if (!card.owner.equals(req.user._id)) {
         res.send({ message: 'Невозможно удалить чужую карточку' });
       } else {
         Card.findByIdAndRemove(req.params.id)
-          .then(() => res.status(200).send({ data: card, message: 'Карточка успешно удалена с сервера' }))
-          .catch(() => res.status(500).send({ message: 'Произошла ошибка' }));
+          .then(() => res.status(CodeSuccess.OK).send({ data: card, message: 'Карточка успешно удалена с сервера' }))
+          .catch((err) => ((err.name === 'CastError')
+            ? res.status(CodeError.BAD_REQUEST).send({ message: 'Переданы некорректные данные при удаления карточки' })
+            : res.status(CodeError.SERVER_ERROR).send({ message: 'Произошла ошибка' })));
       }
     })
-    .catch(() => res.status(500).send({ message: 'Произошла ошибка' }));
+    .catch(() => res.status(CodeError.SERVER_ERROR).send({ message: 'Произошла ошибка' }));
 };
 
-module.exports.likeCard = (req, res, next) => {
+module.exports.likeCard = (req, res) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $addToSet: { likes: req.user._id } },
     { new: true },
   )
-    .then((selectedCard) =>
-      (selectedCard ? res.status(200).send({ data: selectedCard }) : res.status(404).send({ message: 'Передан несуществующий _id карточки' })))
-    .catch((err) => ((err  instanceof CastError) ? res.status(400).send({ message: 'Переданы некорректные данные для постановки/снятии лайка' }) : res.status(500).send({ message: 'Произошла ошибка' })));
+    .then((selectedCard) => (selectedCard
+      ? res.status(CodeSuccess.OK).send({ data: selectedCard })
+      : res.status(CodeError.NOT_FOUND).send({ message: 'Передан несуществующий _id карточки' })))
+    .catch((err) => ((err.name === 'CastError')
+      ? res.status(CodeError.BAD_REQUEST).send({ message: 'Переданы некорректные данные для постановки/снятии лайка' })
+      : res.status(CodeError.SERVER_ERROR).send({ message: 'Произошла ошибка' })));
 };
 
-module.exports.removeLikeCard = (req, res, next) => {
+module.exports.removeLikeCard = (req, res) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $pull: { likes: req.user._id } },
     { new: true },
   )
-  .then((selectedCard) =>
-     (selectedCard ? res.status(200).send({ data: selectedCard }) : res.status(404).send({ message: 'Передан несуществующий _id карточки' })))
-  .catch((err) => ((err instanceof CastError) ? res.status(400).send({ message: 'Переданы некорректные данные для постановки/снятии лайка' }) : res.status(500).send({ message: 'Произошла ошибка' })));
+    .then((selectedCard) => (selectedCard
+      ? res.status(CodeSuccess.OK).send({ data: selectedCard })
+      : res.status(CodeError.NOT_FOUND).send({ message: 'Передан несуществующий _id карточки' })))
+    .catch((err) => ((err.name === 'CastError')
+      ? res.status(CodeError.BAD_REQUEST).send({ message: 'Переданы некорректные данные для постановки/снятии лайка' })
+      : res.status(CodeError.SERVER_ERROR).send({ message: 'Произошла ошибка' })));
 };
